Agrego autenticación http basic

+++ workshop 2/crud-with-rest/app/Http/Controllers/StudentController.php
@@ -9,23 +9,26 @@ class StudentController extends Controller
 {
     public function index() {
         $students = Student::all();
-        return $students;
+        return response()->json($students, 200);
     }
 
     public function create(Request $req) {
         $student = Student::create($req->all());
-        return $student;
+        return response()->json($student, 201);
     }
 
     public function update($id, Request $req) {
         $student = Student::find($id);
         $student->fill($req->all())->save();
-        return $student;
+        return response()->json($student, 200);
     }
 
     public function delete($id) {
         $student = Student::find($id);
-        $student->delete();
-        return $student;
+        if ($student) {
+            $student->delete();
+            return response()->json(null, 204);
+        }
+        return response()->json(null, 500);
     }
 }
